+++ lib/xnix/compatibility/dbus_runtime_client.rb
@@ -65,6 +65,10 @@ module Xnix
         parse_dictionary(call("GetTestPlan", application_id, test_type))
       end
 
+      def test_result(application_id, test_type = "preflight")
+        parse_dictionary(call("GetTestResult", application_id, test_type))
+      end
+
       def snapshot_plan(application_id, reason)
         parse_dictionary(call("GetSnapshotPlan", application_id, reason))
       end